Workaround applied to support embedding on stats sites
The current workaround is based off an AWS serverless architechture that takes care of retrieving and reserving (with appropriate cors settings) a JSON file delivered by an external stats server.

Right now, this is hardcoded and ugly. But it is a first step. I will build a more complete API to take care of setting it up and maintaining it in a more sensible way.

This architechture will likely use a DynamoDB database to store the associations between the stored maps and their JSON urls.