Vendored leaflet-textpath and updated leaflet in bower
The front is drawn again, but the map is broken. Apparently, there is an issue with coordinates (the icons are drawn consistently, then (lower on the screen) the lower half of the map, then (even lower) the upper half of the map.

I also get the error L.drawLocal.draw.handlers.circlemarker is undefined